Don't generate "deb-src" for the target
Our self-built repos don't contain source packages
as well as corresponding "Sources" files and w/o them
"apt" exits with error since it cannot find what was expected.

So for simplicity we exclude all the mess with "deb-src" and
IMHO that's perfectly fine as this tool's functionality is
in building a Debian file-system from existing binary packages
instead of building anything from source.